refact: Rename MainPanel to MainPage
- MainPanel.jsx renamed to MainPage.jsx for structure consistency
